backend/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
defusedxml==0.7.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.3.0
flake8-isort==6.1.2
gunicorn==23.0.0
idna==3.10
isort==6.1.0
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.10
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
shortuuid==1.0.13
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
